\name{SEG3}
\docType{data}
\alias{SEG3}
\title{
ChIP-seq data of transcription factor Ste12 in yeast \emph{S. cerevisiae}
}
\description{
This data set contains parts of ChIP-seq data of transcription factor Ste12 in yeast \emph{S. cerevisiae}. 
ChIP and control data in chromosome 1-3 are kept.
}
\usage{
data("SEG3")
}
\format{
SEG3 has two elements:
\itemize{
\item \code{chip.data} ChIP data in MCS format
\item \code{input.data} control data in MCS format
}
}
\references{
Zheng et al. (2010), \emph{Genetic analysis of variation in transcription factor binding in yeast}, Nature, 464, 1187-1191.
}
